Added autocapitalize="off" to search box.
Added autocapitalize="off" to search box to prevent iOS from auto-capitalizing first char. in the search box on input. I personally find it annoying (using iOS myself) that I have to tap shift to not write uppercase (only first char), instead of just starting to write, right away.

It may be that I'm the only one who feel it is sort of necessary to write all lowercase (by default). It might be the case that I'm bias because of my OCD. :P